requirements.txt
django-authtools==1.6.0
